% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggreg.colors}
\alias{aggreg.colors}
\title{Sort a collection of colors}
\usage{
aggreg.colors(colors, aggreg.method = "median")
}
\arguments{
\item{colors}{a vector containing the colors to be sorted in rgb format.}

\item{aggreg.method}{method of aggregation of the colors corresponding to each word in the \code{sentence} argument. Possible choices are:
\itemize{
   \item \code{most-prominent}: take the first value returned by the function \code{\link{get.color}};
   \item \code{median}: sort all the colors returned by the function \code{\link{get.color}} according to their hue and then take the median color;
   \item \code{mean}: take the mean of all the colors returned by the function \code{\link{get.color}} in the hsv space.
 }}
}
\value{
the 20 first colors found on picular.co for \code{word}. See \url{http://picular.co} for more details
}
\description{
Sort a collection of colors
}
\examples{
get.color("spring")
get.color("fire")
}
\seealso{
\code{\link{create.palette}}
}
