% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregate.colors}
\alias{aggregate.colors}
\title{Parse colors from Picular.co}
\usage{
\method{aggregate}{colors}(colors, aggregate = "median")
}
\arguments{
\item{word}{a string to be converted into a color.}
}
\value{
the 20 first colors found on picular.co for \code{word}. See \url{http://picular.co} for more details
}
\description{
Parse colors from Picular.co
}
\examples{
get.color("spring")
get.color("fire")
}
\seealso{
\code{\link{create.palette}}
}
