% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{create.palette}
\alias{create.palette}
\title{Create a palette given a collection of words}
\usage{
create.palette(words, n.colors = length(words),
  aggregate = "most-prominent")
}
\arguments{
\item{words}{a vector containing}

\item{n.colors}{number of colors of the palette}

\item{aggregate}{method of aggregation}
}
\value{
a color palette of \code{n.colors} elements.
}
\description{
Create a palette given a collection of words
}
\examples{
library(rainbaud)
foo = c("forest", "water", "fire")
pal1 = create.palette(foo)
pal2 = create.palette(foo, n.colors = 11, aggregate = "median")
}
\seealso{
\code{\link{get.color}}
}
