% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbaud-data.R
\docType{data}
\name{drunkenboat}
\alias{drunkenboat}
\title{The Drunken Boat}
\format{An vector of class \code{"character"} containing the text of the poem.}
\usage{
data(drunkenboat)
}
\description{
The Drunken Boat ("Le Bateau ivre") is a poem written by the then 17 years old
 Arthur Rimbaud in 1871 (here provided in the english traslation of Wallace
 Fowlie). The verses were inspired by the recently published "Twenty Thousand
 Leagues Under the Sea" and were part of the letter Rimbaud wrote to introduce
 himself to Paul Verlaine.
}
\examples{
data(drunkenboat)
dbpalette <- create.palette(drunkenboat, n.colors = 5)
plot(1:5, rep(1,5), col = dbpalette, cex = 7, pch = 15, axes = FALSE,
xlab = "", ylab = "", main = "The Drunken Boat")
}
\references{
A. Rimbaud, Complete Works, Selected Letters (The University of Chicago Press, 2005)
(\href{https://www.poetryfoundation.org/poems/55036/the-drunken-boat}{full-text})
}
\keyword{datasets}
