% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create.palette}
\alias{create.palette}
\title{Create a palette given a collection of words (or only one word?)}
\usage{
create.palette(sentence, n.colors = length(words), sorted = TRUE,
  aggregate = "most-prominent")
}
\arguments{
\item{sentence}{a string of words}

\item{n.colors}{number of colors of the palette}

\item{aggregate}{method of aggregation}
}
\value{
a color palette of \code{n.colors} elements.
}
\description{
Create a palette given a collection of words (or only one word?)
}
\examples{
mfl = "the rain in spain stays mainly in the plain"
mfl1 = create.palette(mfl, n.colors = 11)
mfl2 = create.palette(mfl, n.colors = 11, aggregate = "median")
plot(1:11, rep(0, 11), pch = 15, col = mfl1, cex = 4, axes = FALSE, xlab = "", ylab = "")
points(1:11, rep(-0.2, 11), pch = 15, col = mfl1, cex = 4)
}
\seealso{
\code{\link{get.color}}
}
