% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.color}
\alias{get.color}
\title{Parse colors from Picular.co}
\usage{
get.color(word, sorted = TRUE)
}
\arguments{
\item{word}{a string to be converted into a color.}

\item{sorted}{logical, if \code{TRUE} then colors are sorted according to their hue.}
}
\value{
the 20 first colors found on picular.co for \code{word}. See \url{http://picular.co} for more details
}
\description{
Parse colors from Picular.co
}
\examples{
get.color("spring")
get.color("fire")
}
\seealso{
\code{\link{create.palette}}
}
